feat: EvalError canvas

+++ gui/husky-tracer-gui-sycamore/src/components/figure_view/figure_canvas/generic_f32.rs
@@ -3,7 +3,7 @@ use super::*;
 #[derive(Prop)]
 pub struct GenericF32Props<'a> {
     dimension: &'a ReadSignal<PixelDimension>,
-    partitioned_samples: &'a [(PartitionDefnData, Vec<(SampleId, i32)>)],
+    partitioned_samples: &'a [(PartitionDefnData, Vec<(SampleId, f32)>)],
 }
 
 #[component]
